c307: extract TopicLabelTab + share DocsPerTopicBar to workspace/components (#441 2.1)
Workspace.tsx 5544 LOC -> 5183 LOC (-361 LOC). Entry chunk 143.26 KB -> 133.37 KB (-9.9 KB).
25 tabs extracted + lazy.
Moved DocsPerTopicBar helper to workspace/components/DocsPerTopicBar.tsx (still used by inline TopicsTab via import path going forward; in this PR no Workspace.tsx callsite remains but the shared module is in place for c311).
Removed newly-unused TopicLabelHeatmap + InverseLabelHeatmap + DocsPerTopicBar imports from Workspace.tsx.
typecheck clean, 44/44 tests pass.
